[chore] #5 멀티 모듈 패키지 변경
