+++ api/src/main/java/org/sopt/api/common/SuccessStatus.java
@@ -1,0 +1,16 @@
+package org.sopt.api.common;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
+@Getter
+public enum SuccessStatus {
+    OK(HttpStatus.OK, "요청이 성공했습니다."),
+    CREATED(HttpStatus.CREATED, "요청이 성공했습니다.");
+
+    private final HttpStatus httpStatus;
+    private final String message;
+}
