+++ api/src/main/java/org/sopt/api/error/GlobalExceptionHandler.java
@@ -1,0 +1,32 @@
+package org.sopt.api.error;
+
+import lombok.extern.slf4j.Slf4j;
+import org.sopt.api.common.ApiResponse;
+import org.sopt.common.error.BusinessException;
+import org.sopt.common.error.ErrorStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.HttpRequestMethodNotSupportedException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@Slf4j
+@ControllerAdvice
+public class GlobalExceptionHandler {
+    @ExceptionHandler(HttpRequestMethodNotSupportedException.class)
+    protected ResponseEntity<ApiResponse<?>> handleHttpRequestMethodNotSupportedException(final HttpRequestMethodNotSupportedException e) {
+        log.error(">>> handle: HttpRequestMethodNotSupportedException ", e);
+        return ApiResponse.failure(ErrorStatus.METHOD_NOT_ALLOWED);
+    }
+
+    @ExceptionHandler(BusinessException.class)
+    protected ResponseEntity<ApiResponse<?>> handleBusinessException(final BusinessException e) {
+        log.error(">>> handle: BusinessException ", e);
+        return ApiResponse.failure(e.getErrorStatus());
+    }
+
+    @ExceptionHandler(Exception.class)
+    protected ResponseEntity<ApiResponse<?>> handleException(final Exception e) {
+        log.error(">>> handle: Exception ", e);
+        return ApiResponse.failure(ErrorStatus.INTERNAL_SERVER_ERROR);
+    }
+}
