+++ api/src/main/java/org/sopt/api/member/dto/request/MemberSaveRequest.java
@@ -1,0 +1,10 @@
+package org.sopt.api.member.dto.request;
+
+import org.sopt.domain.member.domain.Sopt;
+
+public record MemberSaveRequest(
+        String name,
+        String nickname,
+        int age,
+        Sopt sopt) {
+}
