+++ api/src/main/java/org/sopt/api/member/dto/request/MemberUpdateRequest.java
@@ -1,0 +1,8 @@
+package org.sopt.api.member.dto.request;
+
+import org.sopt.domain.member.domain.Part;
+
+public record MemberUpdateRequest(
+        int generation,
+        Part part) {
+}
