+++ api/src/main/java/org/sopt/api/post/dto/request/PostSaveOrUpdateRequest.java
@@ -1,0 +1,8 @@
+package org.sopt.api.post.dto.request;
+
+public record PostSaveOrUpdateRequest(
+        String title,
+        String postContent,
+        String categoryContent
+) {
+}
