+++ api/src/main/java/org/sopt/api/post/dto/response/PostSaveResponse.java
@@ -1,0 +1,11 @@
+package org.sopt.api.post.dto.response;
+
+import org.sopt.domain.post.domain.Post;
+
+public record PostSaveResponse(
+        Long postId
+) {
+    public static PostSaveResponse of(Post post) {
+        return new PostSaveResponse(post.getId());
+    }
+}
