--- api/src/main/java/org/sopt/member/dto/request/MemberSaveRequest.java
@@ -1,10 +1,0 @@
-package org.sopt.member.dto.request;
-
-import org.sopt.member.domain.Sopt;
-
-public record MemberSaveRequest(
-        String name,
-        String nickname,
-        int age,
-        Sopt sopt) {
-}
