--- api/src/main/java/org/sopt/member/dto/request/MemberUpdateRequest.java
@@ -1,8 +1,0 @@
-package org.sopt.member.dto.request;
-
-import org.sopt.member.domain.Part;
-
-public record MemberUpdateRequest(
-        int generation,
-        Part part) {
-}
