--- api/src/main/java/org/sopt/member/service/MemberService.java
@@ -1,65 +1,0 @@
-package org.sopt.member.service;
-
-import lombok.RequiredArgsConstructor;
-import org.sopt.error.ConflictException;
-import org.sopt.error.ErrorStatus;
-import org.sopt.member.domain.Member;
-import org.sopt.member.domain.Sopt;
-import org.sopt.member.dto.request.MemberSaveRequest;
-import org.sopt.member.dto.request.MemberUpdateRequest;
-import org.sopt.member.dto.response.MemberGetResponse;
-import org.sopt.member.dto.response.MemberSaveResponse;
-import org.sopt.member.repository.MemberRepository;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-import static org.sopt.member.domain.Member.createMember;
-
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-@Service
-public class MemberService {
-    private final MemberRepository memberRepository;
-
-    @Transactional
-    public MemberSaveResponse saveMember(MemberSaveRequest memberSaveRequest) {
-        validateDuplicateMember(memberSaveRequest.nickname());
-        Member member = createMember(memberSaveRequest.name(), memberSaveRequest.nickname(),
-                memberSaveRequest.age(), memberSaveRequest.sopt());
-        Member savedMember = memberRepository.save(member);
-        return MemberSaveResponse.of(savedMember);
-    }
-
-    public MemberGetResponse getMember(Long memberId) {
-        Member findMember = memberRepository.findByIdOrThrow(memberId);
-        return MemberGetResponse.of(findMember);
-    }
-
-    public List<MemberGetResponse> getMembers(Pageable pageable) {
-        return memberRepository.findAll(pageable)
-                .stream()
-                .map(MemberGetResponse::of)
-                .toList();
-    }
-
-    @Transactional
-    public void updateMember(Long memberId, MemberUpdateRequest memberUpdateRequest) {
-        Member findMember = memberRepository.findByIdOrThrow(memberId);
-        Sopt sopt = Sopt.createSopt(memberUpdateRequest.generation(), memberUpdateRequest.part());
-        findMember.updateSopt(sopt);
-    }
-
-    @Transactional
-    public void deleteMember(Long memberId) {
-        memberRepository.deleteById(memberId);
-    }
-
-    private void validateDuplicateMember(String nickname) {
-        if (memberRepository.existsByNickname(nickname)) {
-            throw new ConflictException(ErrorStatus.DUPLICATE_MEMBER);
-        }
-    }
-}
