--- api/src/main/java/org/sopt/post/dto/request/PostSaveOrUpdateRequest.java
@@ -1,8 +1,0 @@
-package org.sopt.post.dto.request;
-
-public record PostSaveOrUpdateRequest(
-        String title,
-        String postContent,
-        String categoryContent
-) {
-}
