--- api/src/main/java/org/sopt/post/dto/response/PostSaveResponse.java
@@ -1,11 +1,0 @@
-package org.sopt.post.dto.response;
-
-import org.sopt.post.domain.Post;
-
-public record PostSaveResponse(
-        Long postId
-) {
-    public static PostSaveResponse of(Post post) {
-        return new PostSaveResponse(post.getId());
-    }
-}
