+++ common/src/main/java/org/sopt/common/error/BusinessException.java
@@ -1,0 +1,13 @@
+package org.sopt.common.error;
+
+import lombok.Getter;
+
+@Getter
+public class BusinessException extends RuntimeException {
+    private final ErrorStatus errorStatus;
+
+    public BusinessException(ErrorStatus errorStatus) {
+        super(errorStatus.getMessage());
+        this.errorStatus = errorStatus;
+    }
+}
