+++ common/src/main/java/org/sopt/common/error/ConflictException.java
@@ -1,0 +1,11 @@
+package org.sopt.common.error;
+
+public class ConflictException extends BusinessException {
+    public ConflictException() {
+        super(ErrorStatus.CONFLICT);
+    }
+
+    public ConflictException(ErrorStatus errorStatus) {
+        super(errorStatus);
+    }
+}
