+++ common/src/main/java/org/sopt/common/error/EntityNotFoundException.java
@@ -1,0 +1,11 @@
+package org.sopt.common.error;
+
+public class EntityNotFoundException extends BusinessException {
+    public EntityNotFoundException() {
+        super(ErrorStatus.ENTITY_NOT_FOUND);
+    }
+
+    public EntityNotFoundException(ErrorStatus errorStatus) {
+        super(errorStatus);
+    }
+}
