+++ common/src/main/java/org/sopt/common/error/InvalidValueException.java
@@ -1,0 +1,11 @@
+package org.sopt.common.error;
+
+public class InvalidValueException extends BusinessException {
+    public InvalidValueException() {
+        super(ErrorStatus.BAD_REQUEST);
+    }
+
+    public InvalidValueException(ErrorStatus errorStatus) {
+        super(errorStatus);
+    }
+}
