--- common/src/main/java/org/sopt/error/BusinessException.java
@@ -1,13 +1,0 @@
-package org.sopt.error;
-
-import lombok.Getter;
-
-@Getter
-public class BusinessException extends RuntimeException {
-    private final ErrorStatus errorStatus;
-
-    public BusinessException(ErrorStatus errorStatus) {
-        super(errorStatus.getMessage());
-        this.errorStatus = errorStatus;
-    }
-}
