--- common/src/main/java/org/sopt/error/ConflictException.java
@@ -1,11 +1,0 @@
-package org.sopt.error;
-
-public class ConflictException extends BusinessException {
-    public ConflictException() {
-        super(ErrorStatus.CONFLICT);
-    }
-
-    public ConflictException(ErrorStatus errorStatus) {
-        super(errorStatus);
-    }
-}
