--- common/src/main/java/org/sopt/error/EntityNotFoundException.java
@@ -1,11 +1,0 @@
-package org.sopt.error;
-
-public class EntityNotFoundException extends BusinessException {
-    public EntityNotFoundException() {
-        super(ErrorStatus.ENTITY_NOT_FOUND);
-    }
-
-    public EntityNotFoundException(ErrorStatus errorStatus) {
-        super(errorStatus);
-    }
-}
