--- common/src/main/java/org/sopt/error/ErrorStatus.java
@@ -1,41 +1,0 @@
-package org.sopt.error;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-
-@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-public enum ErrorStatus {
-    /**
-     * 400 Bad Request
-     */
-    BAD_REQUEST(HttpStatus.BAD_REQUEST, "잘못된 요청입니다."),
-
-    /**
-     * 404 Not Found
-     */
-    ENTITY_NOT_FOUND(HttpStatus.NOT_FOUND, "대상을 찾을 수 없습니다."),
-    MEMBER_NOT_FOUND(HttpStatus.NOT_FOUND, "회원을 찾을 수 없습니다."),
-    POST_NOT_FOUND(HttpStatus.NOT_FOUND, "게시물을 찾을 수 없습니다."),
-
-    /**
-     * 405 Method Not Allowed
-     */
-    METHOD_NOT_ALLOWED(HttpStatus.METHOD_NOT_ALLOWED, "잘못된 HTTP method 요청입니다."),
-
-    /**
-     * 409 Conflict
-     */
-    CONFLICT(HttpStatus.CONFLICT, "이미 존재하는 리소스입니다."),
-    DUPLICATE_MEMBER(HttpStatus.CONFLICT, "이미 존재하는 회원입니다."),
-
-    /**
-     * 500 Internal Server Error
-     */
-    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "서버 내부 오류입니다.");
-
-    private final HttpStatus httpStatus;
-    private final String message;
-}
