--- common/src/main/java/org/sopt/error/InvalidValueException.java
@@ -1,11 +1,0 @@
-package org.sopt.error;
-
-public class InvalidValueException extends BusinessException {
-    public InvalidValueException() {
-        super(ErrorStatus.BAD_REQUEST);
-    }
-
-    public InvalidValueException(ErrorStatus errorStatus) {
-        super(errorStatus);
-    }
-}
