--- domain/src/main/java/org/sopt/config/JpaConfig.java
@@ -1,9 +1,0 @@
-package org.sopt.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@Configuration
-public class JpaConfig {
-}
