+++ domain/src/main/java/org/sopt/domain/config/JpaConfig.java
@@ -1,0 +1,9 @@
+package org.sopt.domain.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@Configuration
+public class JpaConfig {
+}
