+++ domain/src/main/java/org/sopt/domain/member/domain/Member.java
@@ -1,0 +1,52 @@
+package org.sopt.domain.member.domain;
+
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.DynamicUpdate;
+import org.sopt.domain.common.BaseTimeEntity;
+import org.sopt.domain.post.domain.Post;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Builder(access = AccessLevel.PRIVATE)
+@DynamicUpdate
+@Getter
+@Entity
+public class Member extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "member_id")
+    private Long id;
+    private String name;
+    private String nickname;
+    private Integer age;
+    @Embedded
+    private Sopt sopt;
+    @OneToMany(mappedBy = "member", cascade = CascadeType.REMOVE)
+    @Builder.Default
+    private List<Post> posts = new ArrayList<>();
+
+    public static Member createMember(final String name, final String nickname, final Integer age, final Sopt sopt) {
+        return Member.builder()
+                .name(name)
+                .nickname(nickname)
+                .age(age)
+                .sopt(sopt)
+                .build();
+    }
+
+    public void addPost(Post post) {
+        posts.add(post);
+    }
+
+    public void removePost(Post post) {
+        posts.remove(post);
+    }
+
+    public void updateSopt(Sopt sopt) {
+        this.sopt = sopt;
+    }
+}
