--- domain/src/main/java/org/sopt/member/domain/Part.java
@@ -1,17 +1,0 @@
-package org.sopt.member.domain;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-public enum Part {
-    ANDROID("ANDROID"),
-    iOS("iOS"),
-    SERVER("SERVER"),
-    DESIGN("DESIGN"),
-    PLAN("PLAN");
-
-    private final String name;
-}
