--- domain/src/main/java/org/sopt/post/repository/CategoryRepository.java
@@ -1,7 +1,0 @@
-package org.sopt.post.repository;
-
-import org.sopt.post.domain.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Short> {
-}
